food-waste-llm/requirements.txt
pandas
numpy
openai
python-dotenv
matplotlib
plotly
streamlit
statsmodels
